+++ CHANGELOG.md
@@ -1,8 +1,10 @@
 ## [Unreleased]
 
+- Fix replica connection pool getter when database configurations have multiple replicas
+
 ## [0.1.2] - 2024-12-16
 
-Fix CTE regex matcher (4b1d10b)
+- Fix CTE regex matcher (4b1d10b)
 
 ## [0.1.1] - 2024-11-27